feat: playlog as base class
