+++ simaple/app/domain/simulator.py
@@ -6,8 +6,8 @@ from typing import Optional
 
 import pydantic
 
-from simaple.simulate.interface.simulator_configuration import SimulatorConfiguration
 from simaple.simulate.engine import OperationEngine
+from simaple.simulate.interface.simulator_configuration import SimulatorConfiguration
 from simaple.simulate.policy.base import Operation
 from simaple.simulate.policy.parser import parse_dsl_to_operations
 from simaple.simulate.report.dpm import DamageCalculator
