+++ simaple/simulate/reserved_names.py
@@ -1,3 +1,7 @@
+"""
+Pre-defined Constants and Enums for simulation.
+"""
+
 class Tag:
     """Reserved constants for event tag"""
